ORM Completed.

+++ project/src/main/java/com/thinkful/spring/dao/VehicleModelDao.java
@@ -2,8 +2,6 @@ package com.thinkful.spring.dao;
 
 import com.thinkful.spring.entity.VehicleModel;
 
-/**
- * Created by neville on 29/04/15.
- */
-public interface VehicleModelDao extends CrudDao<VehicleModel,Long> {
+
+public interface VehicleModelDao extends CrudDao<VehicleModel,Integer> {
 }
